isolate test statistics logic from tester
